test adaption for travis

+++ tests/testthat/test-deleteAllEntries.R
@@ -1,37 +1,42 @@
 context("deleteAllEntries")
-source("~/.inwt/rsync/config.R")
+expectTrue <- function(a) testthat::expect_true(a)
 
-test_that("deleteAllEntries is working", {
+dirName <- tempdir()
+# invisible(file.remove(paste0(dirName,'/extraFolder')))
+dirName2 <- paste0(dirName, '/extraFolder/')
+dir.create(dirName2)
+# In case we run these Tests multiple times in a row:
+file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
+file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
+x <- 1
+y <- 2
+save(list = "x", file = paste0(dirName, "/", "x.Rdata"))
+save(list = "y", file = paste0(dirName, "/", "y.Rdata"))
 
-  expectTrue <- function(a) testthat::expect_true(a)
+serverTestingRsyncL <- newRsync(from = dirName,
+                                to = dirName2)
 
-  dirName <- tempdir()
-  # invisible(file.remove(paste0(dirName,'/extraFolder')))
-  dirName2 <- paste0(dirName, '/extraFolder/')
-  dir.create(dirName2)
-  # In case we run these Tests multiple times in a row:
-  file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
-  file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
-  x <- 1
-  y <- 2
-  save(list = "x", file = paste0(dirName, "/", "x.Rdata"))
-  save(list = "y", file = paste0(dirName, "/", "y.Rdata"))
 
-  serverTestingRsyncD <- newRsync(from = dirName,
-                                         host = hostURL,
-                                         name = nameServer,
-                                         password = passwordServer)
+if (!is.null(source("~/.inwt/rsync/config.R"))) {
 
-  serverTestingRsyncL <- newRsync(from = dirName,
-                                         to = dirName2)
+  test_that("deleteAllEntries for rsyncD is working", {
 
-  #rsyncD
-  invisible(sendFile(db = serverTestingRsyncD, fileName = 'x.Rdata'))
-  invisible(sendFile(db = serverTestingRsyncD, fileName = 'y.Rdata'))
-  expectTrue(!is.null(listEntries(serverTestingRsyncD)))
-  invisible(deleteAllEntries(db = serverTestingRsyncD))
-  expectTrue(nrow(listEntries(serverTestingRsyncD)) == 0)
+    serverTestingRsyncD <- newRsync(from = dirName,
+                                    host = hostURL,
+                                    name = nameServer,
+                                    password = passwordServer)
 
+    #rsyncD
+    invisible(sendFile(db = serverTestingRsyncD, fileName = 'x.Rdata'))
+    invisible(sendFile(db = serverTestingRsyncD, fileName = 'y.Rdata'))
+    expectTrue(!is.null(listEntries(serverTestingRsyncD)))
+    invisible(deleteAllEntries(db = serverTestingRsyncD))
+    expectTrue(nrow(listEntries(serverTestingRsyncD)) == 0)
+
+  })
+}
+
+test_that("deleteAllEntries for rsyncL is working", {
 
   # rsyncL
   invisible(sendFile(db = serverTestingRsyncL, fileName = 'x.Rdata'))
@@ -39,6 +44,6 @@ test_that("deleteAllEntries is working", {
   expectTrue(!is.null(listEntries(serverTestingRsyncL)))
   invisible(deleteAllEntries(db = serverTestingRsyncL))
   expectTrue(nrow(listEntries(serverTestingRsyncL)) == 0)
-
 })
+
