+++ tests/testthat/test-getEntry.R
@@ -1,39 +1,41 @@
 context("getEntry")
-source("~/.inwt/rsync/config.R")
+expectTrue <- function(a) testthat::expect_true(a)
 
-test_that("getEntry is working", {
+dirName <- tempdir()
+try(file.remove(paste0(dirName,'/extraFolder')), silent = TRUE)
+dirName2 <- paste0(dirName, '/extraFolder/')
+dir.create(dirName2)
+# In case we run these Tests multiple times in a row:
+file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
+file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
+x <- 1
+y <- 2
+save(list = "x", file = paste0(dirName, "/", "x.Rdata"))
+save(list = "y", file = paste0(dirName, "/", "y.Rdata"))
 
-  expectTrue <- function(a) testthat::expect_true(a)
+serverTestingRsyncL <- newRsync(from = dirName,
+                                to = dirName2)
 
+if (!is.null(source("~/.inwt/rsync/config.R"))) {
 
-  dirName <- tempdir()
-  try(file.remove(paste0(dirName,'/extraFolder')), silent = TRUE)
-  dirName2 <- paste0(dirName, '/extraFolder/')
-  dir.create(dirName2)
-  # In case we run these Tests multiple times in a row:
-  file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
-  file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
-  x <- 1
-  y <- 2
-  save(list = "x", file = paste0(dirName, "/", "x.Rdata"))
-  save(list = "y", file = paste0(dirName, "/", "y.Rdata"))
+  test_that("getEntry for rsyncD is working", {
 
-  serverTestingRsyncD <- newRsync(from = dirName,
-                                         host = hostURL,
-                                         name = nameServer,
-                                         password = passwordServer)
+    serverTestingRsyncD <- newRsync(from = dirName,
+                                    host = hostURL,
+                                    name = nameServer,
+                                    password = passwordServer)
 
-  serverTestingRsyncL <- newRsync(from = dirName,
-                                         to = dirName2)
+    # rsyncD
+    expectTrue(nrow(deleteAllEntries(db = serverTestingRsyncD)) == 0)
+    sendFile(db = serverTestingRsyncD, fileName = 'x.Rdata')
+    expectTrue(nrow(listEntries(serverTestingRsyncD)) == 1)
+    file.remove(paste0(dirName, '/','x.Rdata'))
+    getEntry(db = serverTestingRsyncD, entryName = 'x.Rdata')
+    expectTrue(file.exists(paste0(dirName, '/','x.Rdata')))
+  })
+}
 
-
-  # rsyncD
-  expectTrue(nrow(deleteAllEntries(db = serverTestingRsyncD)) == 0)
-  sendFile(db = serverTestingRsyncD, fileName = 'x.Rdata')
-  expectTrue(nrow(listEntries(serverTestingRsyncD)) == 1)
-  file.remove(paste0(dirName, '/','x.Rdata'))
-  getEntry(db = serverTestingRsyncD, entryName = 'x.Rdata')
-  expectTrue(file.exists(paste0(dirName, '/','x.Rdata')))
+test_that("getEntry for rsyncL is working", {
 
   #rsyncL:
   expectTrue(nrow(deleteAllEntries(db = serverTestingRsyncL)) == 0)
@@ -43,4 +45,7 @@ test_that("getEntry is working", {
   getEntry(db = serverTestingRsyncL, entryName = 'x.Rdata')
   expectTrue(file.exists(paste0(dirName, '/','x.Rdata')))
 
+  # #remove traces for further tests
+  file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
+  file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
 })
