+++ tests/testthat/test-listEntries.R
@@ -1,39 +1,38 @@
 context("listEntries")
-source("~/.inwt/rsync/config.R")
+expectTrue <- function(a) testthat::expect_true(a)
 
-test_that("Entries can of rsync objects can be listet", {
+dirName <- tempdir()
+try(file.remove(paste0(dirName,'/extraFolder')), silent = TRUE)
+dirName2 <- paste0(dirName, '/extraFolder/')
+dir.create(dirName2)
+# In case we run these Tests multiple times in a row:
+file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
+file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
+x <- 1
+y <- 2
+save(list = "x", file = paste0(dirName, "/", "x.Rdata"))
+save(list = "y", file = paste0(dirName, "/", "y.Rdata"))
 
-  expectTrue <- function(a) testthat::expect_true(a)
+serverTestingRsyncL <- newRsync(from = dirName,
+                                to = dirName2)
+if (!is.null(source("~/.inwt/rsync/config.R"))) {
+  test_that("Entries rsyncD objects can be listet", {
 
-  dirName <- tempdir()
-  try(file.remove(paste0(dirName,'/extraFolder')), silent = TRUE)
-  dirName2 <- paste0(dirName, '/extraFolder/')
-  dir.create(dirName2)
-  # In case we run these Tests multiple times in a row:
-  file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
-  file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
-  x <- 1
-  y <- 2
-  save(list = "x", file = paste0(dirName, "/", "x.Rdata"))
-  save(list = "y", file = paste0(dirName, "/", "y.Rdata"))
+    serverTestingRsyncD <- newRsync(from = dirName,
+                                    host = hostURL,
+                                    name = nameServer,
+                                    password = passwordServer)
 
+    # rsyncD
+    invisible(deleteAllEntries(db = serverTestingRsyncD))
+    invisible(sendFile(db = serverTestingRsyncD, fileName = 'x.Rdata'))
+    invisible(sendFile(db = serverTestingRsyncD, fileName = 'y.Rdata'))
+    expectTrue(nrow(listEntries(serverTestingRsyncD)) == 2)
+    invisible(deleteAllEntries(db = serverTestingRsyncD))
+  })
+}
 
-
-  serverTestingRsyncD <- newRsync(from = dirName,
-                                         host = hostURL,
-                                         name = nameServer,
-                                         password = passwordServer)
-
-  serverTestingRsyncL <- newRsync(from = dirName,
-                                         to = dirName2)
-
-
-  # rsyncD
-  invisible(deleteAllEntries(db = serverTestingRsyncD))
-  invisible(sendFile(db = serverTestingRsyncD, fileName = 'x.Rdata'))
-  invisible(sendFile(db = serverTestingRsyncD, fileName = 'y.Rdata'))
-  expectTrue(nrow(listEntries(serverTestingRsyncD)) == 2)
-  invisible(deleteAllEntries(db = serverTestingRsyncD))
+test_that("Entries rsyncL objects can be listet", {
 
   # rsyncL
   invisible(deleteAllEntries(db = serverTestingRsyncL))
@@ -42,7 +41,7 @@ test_that("Entries can of rsync objects can be listet", {
   expectTrue(nrow(listEntries(serverTestingRsyncL)) == 2)
   invisible(deleteAllEntries(db = serverTestingRsyncL))
 
-#   #remove traces for further tests
-#   file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
-#   file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
- })
+  #   #remove traces for further tests
+  #   file.remove(dir(dirName2, "Rdata|csv|json", full.names = TRUE))
+  #   file.remove(dir(dirName, "Rdata|csv|json", full.names = TRUE))
+})
